:candle: refactor: Move all reflection types to a new file inside package
- Refactor standard_injector to use named returns
- Add missing test check
