+++ remy/internal/utils/generics.go
@@ -3,8 +3,6 @@ package utils
 import (
 	"fmt"
 	"github.com/wrapped-owls/goremy-di/remy/internal/types"
-	"reflect"
-	"strings"
 )
 
 // nilStr is the default const string representation of a nil type
@@ -35,50 +33,6 @@ func TypeName[T any](elements ...T) (name string) {
 	name = fmt.Sprintf("%T", value)
 	if name == nilStr {
 		name = fmt.Sprintf("%T", &value)
-	}
-	return
-}
-
-// TypeNameByReflect returns a string that defines the name of the given generic type.
-func TypeNameByReflect(generifyInterface bool, elementType reflect.Type, isInterface bool) string {
-	if elementType == nil {
-		panic(ErrImpossibleIdentifyType)
-	}
-
-	if generifyInterface && isInterface {
-		var builder strings.Builder
-		builder.WriteString("interface { ")
-		for i := 0; i < elementType.NumMethod(); i++ {
-			if i > 0 {
-				builder.WriteString("; ")
-			}
-			builder.WriteString(
-				fmt.Sprintf(
-					"%s %s", elementType.Method(i).Name, elementType.Method(i).Type,
-				),
-			)
-		}
-		builder.WriteString(" }")
-		return builder.String()
-	}
-	return fmt.Sprintf("%s/%s{###}%s", elementType.PkgPath(), elementType.Name(), fmt.Sprint(elementType))
-}
-
-func GetElemType(element any) (foundType reflect.Type, isInterface bool) {
-	foundType = reflect.TypeOf(element)
-	if foundType == nil {
-		panic(ErrImpossibleIdentifyType)
-	}
-	return
-}
-
-func GetType[T any]() (foundType reflect.Type, isInterface bool) {
-	var typeT T
-	foundType = reflect.TypeOf(typeT)
-	if foundType == nil {
-		// T is an interface
-		isInterface = true
-		foundType = reflect.TypeOf(&typeT)
 	}
 	return
 }
